# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Swagger SaveTheMeal! - API 1.0",
    "description": "This is a sample meal Server for save the food.\nThat way, with time, we can improve the API in general, and expose some of the features",
    "contat": {
      "name": "Alessia,Anna,Alessio"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Meal",
      "description": "API for meals in the system"
    },
    {
      "name": "Fornitore",
      "description": "API for suppliers in the system"
    },
    {
      "name": "Utente",
      "description": "Operations about user"
    },
    {
      "name": "Feedback",
      "description": "Operations about feedbacks"
    },
    {
      "name": "Acquisto",
      "description": "Operations about purchase"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/meal": {
      "get": {
        "tags": ["Meal"],
        "summary": "Get all meals in system, or all meals filtered by a supplier",
        "description": "",
        "parameters": [
          {
            "name": "fornitore",
            "in": "query",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Meals"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Meal"],
        "summary": "Add a new meal to the store",
        "parameters": [
          {
            "name": "Meal",
            "description": "meal object",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Meals"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Meals"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Meal"],
        "summary": "Delete all meals in system",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Meals"
                }
              }
            }
          }
        }
      }
    },
    "/meal/{codiceID}": {
      "get": {
        "tags": ["Meal"],
        "summary": "Finds Meals by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Meal id",
            "required": true,
            "schema": {
              "tipe": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Meals"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Meal"],
        "summary": "Finds and delete Meals by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Meal id",
            "required": true,
            "schema": {
              "tipe": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Meals"
                }
              }
            }
          }
        }
      }
    },
    "/fornitore": {
      "get": {
        "tags": ["Fornitore"],
        "summary": "Get all suppliers in  the system",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Fornitori"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Fornitore"],
        "summary": "Add a new supplier in the system",
        "description": "",
        "parameters": [
          {
            "name": "Fornitore",
            "description": "fornitore objets",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Fornitori"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Fornitori"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Fornitore"],
        "summary": "Delete all suppliers in  the system",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Fornitori"
                }
              }
            }
          }
        }
      }
    },
    "/fornitore/{id}": {
      "get": {
        "tags": ["Fornitore"],
        "summary": "Finds suppliers by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "fornitore id",
            "required": true,
            "schema": {
              "tipe": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Fornitori"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Fornitore"],
        "summary": "Finds and delete suppliers by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "fornitore id",
            "required": true,
            "schema": {
              "tipe": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Fornitori"
                }
              }
            }
          }
        }
      }
    },
    "/utente": {
      "get": {
        "tags": ["Utente"],
        "summary": "Get all users in  the system",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Utente"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Utente"],
        "summary": "Add a new user in the system",
        "description": "",
        "parameters": [
          {
            "name": "Utente",
            "description": "user object",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Utenti"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Utenti"
                }
              }
            }
          }
        }
      }
    },
    "/feedback": {
      "get": {
        "tags": ["Feedback"],
        "summary": "Get all feedbacks in  the system",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Feedback"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Feedback"],
        "summary": "Add a new feedback in the system",
        "description": "",
        "parameters": [
          {
            "name": "Feedback",
            "description": "feedback object",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Feedbacks"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Feedbacks"
                }
              }
            }
          }
        }
      }
    },
    "/acquisto": {
      "get": {
        "tags": ["Acquisto"],
        "summary": "Get all purchases in  the system",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Acquisti"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Acquisto"],
        "summary": "Add a new purchase in the system",
        "description": "",
        "parameters": [
          {
            "name": "Aquisto",
            "description": "purchase object",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Acquisti"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Acquisti"
                }
              }
            }
          }
        }
      }
    },
    "/acquisto/{id}": {
      "get": {
        "tags": ["Acquisto"],
        "summary": "Finds purchases by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "purchase id",
            "required": true,
            "schema": {
              "tipe": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Acquisti"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Meals": {
      "required": [
        "fornitore",
        "prezzo",
        "dimensione",
        "disponibilita"
      ],
      "properties": {
        "fornitore": {
          "type": "string"
        },
        "prezzo": {
          "type": "number"
        },
        "dimensione": {
          "type": "string"
        },
        "disponibilita": {
          "type": "boolean"
        }
      }
    },
    "Fornitori": {
      "required": [
        "email",
        "password",
        "nomeAttivita",
        "indirizzoNegozio",
        "tipologiaAlimenti",
        "IBAN",
        "immagine"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "nomeAttivita": {
          "type": "string"
        },
        "indirizzoNegozio": {
          "type": "string"
        },
        "tipologiaAlimenti": {
          "type": "string"
        },
        "IBAN": {
          "type": "string"
        },
        "immagine": {
          "type": "string"
        }
      }
    },
    "Feedbacks": {
      "required": [
        "fornitore",
        "utente",
        "valutazione",
        "puntiDiForza",
        "commento"
      ],
      "properties": {
        "fornitore": {
          "type": "string"
        },
        "utente": {
          "type": "string"
        },
        "valutazione": {
          "type": "number"
        },
        "puntiDiForza": {
          "type": "string"
        },
        "commento": {
          "type": "string"
        }
      }
    },
    "Acquisti": {
      "required": [
        "meal",
        "acquirente",
        "presenzaIntolleranze",
        "intolleranze",
        "isPaid",
        "borsa",
        "stato"
      ],
      "properties": {
        "meal": {
          "type": "string"
        },
        "acquirente": {
          "type": "string"
        },
        "presenzaIntolleranze": {
          "type": "boolean"
        },
        "intolleranze": {
          "type": "string"
        },
        "isPaid": {
          "type": "boolean"
        },
        "borsa": {
          "type": "boolean"
        },
        "stato": {
          "type": "string"
        }
      }
    },
    "Utenti": {
      "required": ["email", "password", "nome", "cognome"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "nome": {
          "type": "string"
        },
        "cognome": {
          "type": "string"
        }
      }
    }
  }
}